+++ src/constants/index.js
@@ -1,57 +1,15 @@
-export const PAGE = {
-  CARD_LIST: 'cardList',
-  ADD_CARD_FORM: 'addCardForm',
-  ADD_CARD_COMPLETE: 'addCardComplete',
+import { PAGE } from './page';
+import {
+  CARD_COMPANY_LIST,
+  CARD_NUMBER_UNIT_LENGTH,
+  EXPIRATION_DATE_FORMAT_LENGTH,
+  MAX_OWNER_NAME_LENGTH,
+} from './card';
+
+export {
+  CARD_COMPANY_LIST,
+  CARD_NUMBER_UNIT_LENGTH,
+  EXPIRATION_DATE_FORMAT_LENGTH,
+  MAX_OWNER_NAME_LENGTH,
+  PAGE,
 };
-
-export const CARD_NUMBER_UNIT_LENGTH = 4;
-
-export const CARD_COMPANY_LIST = [
-  {
-    name: '신한카드',
-    color: 'rgb(226, 65, 65)',
-    patterns: ['111111', '111100'],
-  },
-  {
-    name: 'KB국민카드',
-    color: 'rgb(84, 124, 228)',
-    patterns: ['222222', '222200'],
-  },
-  {
-    name: '우리카드',
-    color: 'rgb(115, 188, 109)',
-    patterns: ['3333333', '333300'],
-  },
-  {
-    name: '삼성카드',
-    color: 'rgb(222, 89, 185)',
-    patterns: ['444444', '444400'],
-  },
-  {
-    name: 'NH농협카드',
-    color: 'rgb(148, 218, 205)',
-    patterns: ['555555', '555500'],
-  },
-  {
-    name: '하나카드',
-    color: 'rgb(231, 110, 154)',
-    patterns: ['666666', '666600'],
-  },
-  {
-    name: '현대카드',
-    color: 'rgb(243, 125, 59)',
-    patterns: ['777777', '777700'],
-  },
-  {
-    name: '카카오뱅크',
-    color: ' rgb(251, 205, 88)',
-    patterns: ['888888', '888800'],
-  },
-];
-
-export const EXPIRATION_DATE_FORMAT_LENGTH = {
-  month: 2,
-  year: 2,
-};
-
-export const MAX_OWNER_NAME_LENGTH = 30;
