+++ src/constants/page.js
@@ -1,0 +1,5 @@
+export const PAGE = {
+  CARD_LIST: 'cardList',
+  ADD_CARD_FORM: 'addCardForm',
+  ADD_CARD_COMPLETE: 'addCardComplete',
+};
